requirements.txt
aiofiles==0.4.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrdict==2.0.1
awscli==1.16.82
blis==0.4.1
botocore==1.12.72
catalogue==0.2.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.3.9
cymem==2.0.3
dnspython==1.16.0
email-validator==1.0.5
en-core-web-md==2.2.5
en-core-web-sm==2.2.5
fastapi==0.45.0
future==0.18.2
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.0.13
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
joblib==0.14.1
MarkupSafe==1.1.1
more-itertools==8.0.2
murmurhash==1.0.2
mypy==0.750
mypy-extensions==0.4.1
newsapi-python==0.2.6
numpy==1.17.4
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
promise==2.3
psutil==5.6.7
pyasn1==0.4.5
pydantic==1.3
python-multipart==0.0.5
python-twitter==3.5
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
Rx==1.6.1
s3transfer==0.1.13
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
spacy==2.2.3
srsly==1.0.0
starlette==0.12.9
thinc==7.3.1
tqdm==4.41.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.7
uvicorn==0.11.1
uvloop==0.14.0
wasabi==0.5.0
websockets==8.1
zipp==0.6.0
